Extract LevelLogProcessor module
