+++ lib/line-builders.js
@@ -15,23 +15,6 @@ function appendTime (lineParts, { prettified }) {
   }
 }
 
-function appendLevel (lineParts, { prettified, levelFirst }) {
-  const { prettifiedLevel } = prettified
-  if (prettifiedLevel) {
-    if (levelFirst) {
-      if (lineParts.length > 0) {
-        lineParts.unshift(' ')
-      }
-      lineParts.unshift(prettifiedLevel)
-    } else {
-      if (lineParts.length > 0) {
-        lineParts.push(' ')
-      }
-      lineParts.push(prettifiedLevel)
-    }
-  }
-}
-
 function appendMetadata (lineParts, { prettified }) {
   const { prettifiedMetadata } = prettified
   if (prettifiedMetadata) {
@@ -88,7 +71,6 @@ function appendObjectOrError (lineParts, { log, EOL, IDENT, errorLikeObjectKeys,
 
 const builtInLineBuilders = {
   appendTime,
-  appendLevel,
   appendMetadata,
   appendSemicolon,
   appendMessage,
