+++ lib/log-processors.js
@@ -1,13 +1,13 @@
-const { prettifyLevel, prettifyMessage, prettifyMetadata, prettifyTime } = require('./utils')
+const { prettifyMessage, prettifyMetadata, prettifyTime } = require('./utils')
 const { builtInLineBuilders } = require('./line-builders')
 const LogProcessor = require('./LogProcessor')
 const JsonLogProcessor = require('./processors/JsonLogProcessor')
 const PrimitivesLogProcessor = require('./processors/PrimitivesLogProcessor')
 const SearchLogProcessor = require('./processors/SearchLogProcessor')
 const IgnoreLogProcessor = require('./processors/IgnoreLogProcessor')
+const LevelLogProcessor = require('./processors/LevelLogProcessor')
 
 const {
-  appendLevel,
   appendMessage,
   appendMetadata,
   appendSemicolon,
@@ -28,16 +28,6 @@ function createLogProcessor (definition) {
     return new builtInLogProcessors[definition]()
   } else {
     return definition
-  }
-}
-
-class LevelLogProcessor extends LogProcessor {
-  parse (input, context) {
-    return prettifyLevel(input, context)
-  }
-
-  build (input, context) {
-    return appendLevel(input, context)
   }
 }
