+++ lib/utils.js
@@ -14,7 +14,6 @@ const {
 module.exports = {
   isObject,
   prettifyErrorLog,
-  prettifyLevel,
   prettifyMessage,
   prettifyMetadata,
   prettifyObject,
@@ -159,26 +158,6 @@ function prettifyErrorLog ({
   }
 
   return result
-}
-
-/**
- * Checks if the passed in `input` object has a `level` value and returns a prettified
- * string for that level if so.
- *
- * @param {object} input The log object which should have a `level` property.
- * @param {object} context The `context` object which should have a `colorizer` property.
- * @param {function} [context.colorizer] A colorizer function that accepts a level
- * value and returns a colorized string. Default: a no-op colorizer.
- *
- * @returns {undefined|string} If `input` does not have a `level` property then
- * `undefined` will be returned. Otherwise, a string from the specified
- * `colorizer` is returned.
- */
-function prettifyLevel (input, { prettified, colorizer = defaultColorizer } = {}) {
-  if ('level' in input === false) return undefined
-  const colorized = colorizer(input.level)
-  prettified.prettifiedLevel = colorized
-  return undefined
 }
 
 /**
